move filter service back to global filters location
