+++ src/app/components/filters/age-groups/age-groups.component.ts
@@ -2,7 +2,7 @@ import { Component, CUSTOM_ELEMENTS_SCHEMA, OnInit } from '@angular/core';
 import { Observable, Subscription } from 'rxjs/Rx';
 import { Router } from '@angular/router';
 
-import { FilterService } from '../filter.service';
+import { FilterService } from '../../../services/filter.service';
 import { LookupService } from '../../../services/lookup.service';
 import { AgeGroup } from '../../../models/age-group';
 import { awsFieldNames } from '../../../shared/constants';
