+++ src/app/components/filters/kids-welcome/kids-welcome.component.spec.ts
@@ -4,7 +4,7 @@ import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 import { By } from '@angular/platform-browser';
 import { Observable } from 'rxjs/Rx';
 
-import { FilterService } from '../filter.service';
+import { FilterService } from '../../../services/filter.service';
 import { MockTestData } from '../../../shared/MockTestData';
 import { KidsWelcomeComponent } from './kids-welcome.component';
