+++ src/app/components/filters/online-or-physical-group/online-or-physical-group.component.spec.ts
@@ -6,7 +6,7 @@ import { Observable } from 'rxjs/Rx';
 
 import { OnlineOrPhysicalGroupComponent } from './online-or-physical-group.component';
 
-import { FilterService } from '../filter.service';
+import { FilterService } from '../../../services/filter.service';
 import { StateService } from '../../../services/state.service';
 import { ViewType } from '../../../shared/constants';
